replace "norelocate" with "256M" option
for upper phys address. Relocation is required to avoid overlaps. seL4 kernel
however fails to load if the modules are at the end of the physical memory.
So add a 256M option as upper bound, so that the kernel behind finds free
memory (if you have a machine with less then 256M you're lost of course,
for various reasons).