Revert "Don't lose last character of command line."
This reverts commit f95c44ada85866da5bf96f84a6c93d6e5c268ccc.

This commit caused an ELF parsing problem when booting foc_x86_64 with a
'bootstrap' module with 'modlen' of 0x28ff0 bytes and 'slen' of 0x10 bytes,
which adds up to 0x29000 bytes. The problem was caused by the mismatch between
the size calculation in line 135 (without +1) and the +1 introduced by the
commit. The 'slen' calculation in line 128 already covers the whole string
length with zero-termination, so it is not really clear why the additional +1
of this commit would be needed.